test q4 with float16 & bf16

+++ tests/test_q4_cuda.py
@@ -17,7 +17,7 @@ GENERATE_EVAL_SIZE = 100
 class TestsQ4CUDA(unittest.TestCase):
     @parameterized.expand(
         [
-            (torch.float32, "cuda:0"),
+            (torch.bfloat16, "cuda:0"),
             (torch.float16, "cuda:0"),
         ]
     )
@@ -64,7 +64,7 @@ class TestsQ4CUDA(unittest.TestCase):
 
     @parameterized.expand(
         [
-            (torch.float32, "cuda:0"),
+            (torch.bfloat16, "cuda:0"),
             (torch.float16, "cuda:0"),
         ]
     )
